move terms to core-ui and fix the related stuff
