+++ src/app/core-ui/main/terms/modal/terms.component.html
